Make ethtool-lite behave more sanely.  Closes: #496647
For some reason, ethtool-lite didn't distinguish between "SIOCETHTOOL says
the link is down" and "ethtool failed; try mii instead".  This resulted in
some degree of confusion when drivers like the e1000e don't work "as
expected" with mii.  Now, if SIOCETHTOOL says "yes, I know the link is
down", we trust it.