Clarify the requirements for hostnames.  Closes: #399071
Originally, this bug report was aimed at removing uppercase letters from
hostnames, as well as modifying the documentation to suit.  After some
consideration, though, I figure there are plenty of people with uppercase
letters in hostnames, and if they're happy to deal with the consequences,
why should we worry about it?  It'll save bug reports in wheezy saying "my
hostname is no longer accepted", anyway.