Extend link detection time, with a progress bar.  Closes: #414117
Simple function to check repetitively for link and update a progress bar as
we do so.  In the simplest case, we'll find link immediately and move on
with our lives; in the worst case, the setup will be delayed by 3 seconds
times the number of interfaces (if there is no link anywhere).  It will,
however, not slow anything down if link is detected immediately, and it
should definitely solve the "link detection" problems people have been
seeing.