Protect dhcp sigchld handler from races
The core problem here is that SIGCHLD doesn't tell you which of your many
children may have exited... so when the process spawned by popen() finishes,
it will trigger that SIGCHLD handler, and if it's waiting for a process that
isn't likely to finish any time soon, things get ugly.  So, we'll wait with
WNOHANG, and check the status code from that, to determine whether the
right child process has exited.

Unix process management... <shudder>